+++ docs/playroom/snippets.ts
@@ -955,6 +955,203 @@ const snippets: Array<Snippet> = [
 		</TableWrapper>`,
 	},
 	{
+		group: 'Table',
+		name: 'Sortable columns',
+		code: `<TableWrapper>
+		<Table striped>
+			<TableCaption>
+				Population of Australian states and territories, December 2018
+				<VisuallyHidden>
+					, column headers with buttons are sortable.
+				</VisuallyHidden>
+			</TableCaption>
+			<TableHead>
+				<TableRow>
+					<TableHeaderSortable
+						scope="col"
+						sort="ASC"
+						width="50%"
+						onClick={console.log}
+					>
+						Location
+					</TableHeaderSortable>
+					<TableHeaderSortable
+						scope="col"
+						width="50%"
+						onClick={console.log}
+						textAlign="right"
+					>
+						Population
+					</TableHeaderSortable>
+				</TableRow>
+			</TableHead>
+			<TableBody>
+				<TableRow>
+					<TableCell>Australian Capital Territory</TableCell>
+					<TableCell textAlign="right">393,000</TableCell>
+				</TableRow>
+				<TableRow>
+					<TableCell>New South Wales</TableCell>
+					<TableCell textAlign="right">7,670,700</TableCell>
+				</TableRow>
+				<TableRow>
+					<TableCell>Northern Territory</TableCell>
+					<TableCell textAlign="right">244,400</TableCell>
+				</TableRow>
+				<TableRow>
+					<TableCell>Queensland</TableCell>
+					<TableCell textAlign="right">4,808,800</TableCell>
+				</TableRow>
+				<TableRow>
+					<TableCell>South Australia</TableCell>
+					<TableCell textAlign="right">1,702,800</TableCell>
+				</TableRow>
+				<TableRow>
+					<TableCell>Tasmania</TableCell>
+					<TableCell textAlign="right">517,400</TableCell>
+				</TableRow>
+				<TableRow>
+					<TableCell>Victoria</TableCell>
+					<TableCell textAlign="right">5,996,400</TableCell>
+				</TableRow>
+				<TableRow>
+					<TableCell>Western Australia</TableCell>
+					<TableCell textAlign="right">2,603,900</TableCell>
+				</TableRow>
+			</TableBody>
+		</Table>
+	</TableWrapper>`,
+	},
+	{
+		group: 'Table',
+		name: 'With selectable rows',
+		code: `<Stack>
+			<Box paddingBottom={0.75} paddingLeft={0.75} borderBottom>
+				<Checkbox
+					checked={false}
+					onChange={console.log}
+					size="sm"
+				>
+					Select all rows
+				</Checkbox>
+			</Box>
+				<TableWrapper>
+				<Table>
+					<TableHead>
+						<TableRow>
+							<TableHeader>Select</TableHeader>
+							<TableHeader scope="col">Reference</TableHeader>
+							<TableHeader scope="col">Date submitted</TableHeader>
+							<TableHeader scope="col">Actions</TableHeader>
+						</TableRow>
+					</TableHead>
+					<TableBody>		
+						<TableRow selected={false}>
+							<TableCell>
+								<Checkbox
+									size="sm"
+									checked={false}
+									onChange={console.log}
+								>
+									<VisuallyHidden>Select</VisuallyHidden>
+								</Checkbox>
+							</TableCell>
+							<TableCell as="th" scope="row" fontWeight="bold">
+								<TextLink href="#" id="REF-AB0CD4EF">
+									REF-AB0CD4EF
+								</TextLink>
+							</TableCell>
+							<TableCell>21/06/2024</TableCell>
+							<TableCell>
+								<Flex gap={1}>
+									<TextLink
+										aria-describedby="REF-AB0CD4EF"
+										href="#"
+									>
+										Download
+									</TextLink>
+									<TextLink
+										aria-describedby="REF-AB0CD4EF"
+										href="#"
+									>
+										Delete
+									</TextLink>
+								</Flex>
+							</TableCell>
+						</TableRow>
+						
+						<TableRow selected={false}>
+							<TableCell>
+								<Checkbox
+									size="sm"
+									checked={false}
+									onChange={console.log}
+								>
+									<VisuallyHidden>Select</VisuallyHidden>
+								</Checkbox>
+							</TableCell>
+							<TableCell as="th" scope="row" fontWeight="bold">
+								<TextLink href="#" id="REF-AB1CD4EF">
+									REF-AB1CD4EF
+								</TextLink>
+							</TableCell>
+							<TableCell>22/06/2024</TableCell>
+							<TableCell>
+								<Flex gap={1}>
+									<TextLink
+										aria-describedby="REF-AB1CD4EF"
+										href="#"
+									>
+										Download
+									</TextLink>
+									<TextLink
+										aria-describedby="REF-AB1CD4EF"
+										href="#"
+									>
+										Delete
+									</TextLink>
+								</Flex>
+							</TableCell>
+						</TableRow>
+						<TableRow selected={false}>
+							<TableCell>
+								<Checkbox
+									size="sm"
+									checked={false}
+									onChange={console.log}
+								>
+									<VisuallyHidden>Select</VisuallyHidden>
+								</Checkbox>
+							</TableCell>
+							<TableCell as="th" scope="row" fontWeight="bold">
+								<TextLink href="#" id="REF-AB2CD4EF">
+									REF-AB2CD4EF
+								</TextLink>
+							</TableCell>
+							<TableCell>22/06/2024</TableCell>
+							<TableCell>
+								<Flex gap={1}>
+									<TextLink
+										aria-describedby="REF-AB2CD4EF"
+										href="#"
+									>
+										Download
+									</TextLink>
+									<TextLink
+										aria-describedby="REF-AB2CD4EF"
+										href="#"
+									>
+										Delete
+									</TextLink>
+								</Flex>
+							</TableCell>
+						</TableRow>
+					</TableBody>
+				</Table>
+			</TableWrapper>
+		</Stack>`,
+	},
+	{
 		group: 'DatePicker',
 		name: 'Basic',
 		code: `<DatePicker label="Select date" value={new Date()} onChange={() => {}} />`,
@@ -1137,7 +1334,41 @@ const snippets: Array<Snippet> = [
 	{
 		group: 'Card',
 		name: 'Basic',
+		code: `<Card>
+			<CardInner>
+				<Stack gap={1}>
+				<H3>
+					<CardLink href="#">Card heading</CardLink>
+				</H3>
+				<Text as="p">
+					Lorem ipsum dolor, sit amet consectetur adipisicing elit. In,
+					voluptat
+				</Text>
+				</Stack>
+			</CardInner>
+		</Card>`,
+	},
+	{
+		group: 'Card',
+		name: 'Clickable',
 		code: `<Card shadow clickable>
+			<CardInner>
+				<Stack gap={1}>
+				<H3>
+					<CardLink href="#">Card heading</CardLink>
+				</H3>
+				<Text as="p">
+					Lorem ipsum dolor, sit amet consectetur adipisicing elit. In,
+					voluptat
+				</Text>
+				</Stack>
+			</CardInner>
+		</Card>`,
+	},
+	{
+		group: 'Card',
+		name: 'Clickable with Footer outside',
+		code: `<Card shadow clickable footer={<CardFooter><TextLink href="#">Action</TextLink></CardFooter>} footerOutside>
 			<CardInner>
 				<Stack gap={1}>
 				<H3>
@@ -1321,32 +1552,57 @@ const snippets: Array<Snippet> = [
 	{
 		group: 'SectionAlert',
 		name: 'Success high with description',
-		code: `<SectionAlert tone="successHigh" title="Success"><Text as="p">Lorem ipsum dolor sit amet, consectetur adipiscing elit.</Text></SectionAlert>`,
+		code: `<SectionAlert title="Success" tone="successHigh"><Text as="p">Lorem ipsum dolor sit amet, consectetur adipiscing elit.</Text></SectionAlert>`,
 	},
 	{
 		group: 'SectionAlert',
 		name: 'Success medium',
-		code: `<SectionAlert tone="successMedium" title="Success"></SectionAlert>`,
+		code: `<SectionAlert title="Success" tone="successMedium"></SectionAlert>`,
 	},
 	{
 		group: 'SectionAlert',
 		name: 'Success low',
-		code: `<SectionAlert tone="successLow" title="Success"></SectionAlert>`,
+		code: `<SectionAlert title="Success" tone="successLow"></SectionAlert>`,
 	},
 	{
 		group: 'SectionAlert',
 		name: 'Error high',
-		code: `<SectionAlert tone="errorHigh" title="Error"></SectionAlert>`,
+		code: `<SectionAlert title="Error" tone="errorHigh"></SectionAlert>`,
 	},
 	{
 		group: 'SectionAlert',
 		name: 'Warning high',
-		code: `<SectionAlert tone="warningHigh" title="Warning"></SectionAlert>`,
+		code: `<SectionAlert title="Warning" tone="warningHigh"></SectionAlert>`,
 	},
 	{
 		group: 'SectionAlert',
 		name: 'Info high with dismissable',
-		code: `<SectionAlert tone="infoHigh" title="Info" onClose={console.log}></SectionAlert>`,
+		code: `<SectionAlert onClose={console.log} title="Info" tone="infoHigh"></SectionAlert>`,
+	},
+	{
+		group: 'SectionAlert',
+		name: 'notStartedLow',
+		code: `<SectionAlert title="Cannot start" tone="cannotStartLow"></SectionAlert>`,
+	},
+	{
+		group: 'SectionAlert',
+		name: 'inProgressLow',
+		code: `<SectionAlert title="In progress" tone="inProgressLow"></SectionAlert>`,
+	},
+	{
+		group: 'SectionAlert',
+		name: 'pausedLow',
+		code: `<SectionAlert title="Paused" tone="pausedLow"></SectionAlert>`,
+	},
+	{
+		group: 'SectionAlert',
+		name: 'cannotStartLow',
+		code: `<SectionAlert title="Not started" tone="notStartedLow"></SectionAlert>`,
+	},
+	{
+		group: 'SectionAlert',
+		name: 'unknownLow',
+		code: `<SectionAlert title="Unknown" tone="unknownLow"></SectionAlert>`,
 	},
 	{
 		group: 'PasswordInput',